feat: implement finalize_stream function and split test files
- Add finalize_stream function that executes a finalizer when stream completes
- Supports both sync and async finalizers
- Ensures finalizer is called even if stream raises exceptions
- Split test_aggregation.py into basic and advanced test files to respect 1000-line limit
- test_aggregation_basic.py: tests for batch, chunk, window, scan, group_by, distinct, pairwise, memoize (577 lines)
- test_aggregation_advanced.py: tests for buffer, expand, finalize functions (446 lines)
- Comprehensive test coverage for all finalize_stream scenarios
